# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Royal SL E-commerce API",
        "description": "REST API for the Royal SL Ecommerce System — admin panel, customer storefront, POS, warehouse & inventory management, coupons, payments, and reporting. All sales channels (website, Facebook, outlet POS) share a single central inventory.",
        "contact": {
            "email": "support@royalsl.example"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
...
        "/admin/orders": {
            "get": {
                "tags": [
                    "Admin Orders"
                ],
                "summary": "List/search orders across all channels (website, facebook, pos)",
                "operationId": "97c74ec8e37842a11bfd2279d4f453f0",
                "parameters": [
                    {
                        "name": "source",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "website",
                                "facebook",
                                "pos"
                            ]
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated order list"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/orders/facebook": {
            "post": {
                "tags": [
                    "Admin Orders"
                ],
                "summary": "Import an order received via Facebook / Messenger commerce",
                "operationId": "eb4ee93191ec995a80d6c5fb62c6c0a1",
                "responses": {
                    "201": {
                        "description": "Order created and inventory deducted"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/orders/{order}/status": {
            "put": {
                "tags": [
                    "Admin Orders"
                ],
                "summary": "Update an order status",
                "operationId": "48fe386cae46f9fe330563368893e9c2",
                "responses": {
                    "200": {
                        "description": "Updated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Sanctum personal access token. Send as: Authorization: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}